Fix button component Co-authored-by: Raibrr <ramon.ibrr@gmail.com> Co-authored-by: madera-design <josemadera289@gmail.com> Co-authored-by: luisgarciaes <luisg.garciaespinoza@gmail.com>
