--- FlightReservations/components/Button/styles.tsx
@@ -1,12 +1,0 @@
-import {StyleSheet} from 'react-native';
-
-export const styles = StyleSheet.create({
-  btn__container: {
-    backgroundColor: 'blue',
-  },
-  btn__text: {
-    color: 'white',
-    fontSize: 20,
-    fontWeight: '500',
-  },
-});
